SET-338 Only log unsubscribtion if u ser has been a member of a team / org
